+++ tests/unit/File/ParseExceptionTest.php
@@ -11,12 +11,26 @@ use PHPUnit\Framework\TestCase;
  */
 class ParseExceptionTest extends TestCase
 {
+    /**
+     * @return ParseException
+     */
     public function testCreation()
     {
         $exception = new ParseException('message string');
         $this->assertInstanceOf('Ktomk\Pipelines\File\ParseException', $exception);
         $this->assertInstanceOf('InvalidArgumentException', $exception);
         $this->assertInstanceOf('Exception', $exception);
+
+        return $exception;
+    }
+
+    /**
+     * @param ParseException $exception
+     * @depends testCreation
+     */
+    public function testGetParseMessage(ParseException $exception)
+    {
+        $this->assertNull($exception->getParseMessage());
     }
 
     /**